requirements.txt
Flask==2.0.3
Flask_Login==0.6.0
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.0
python-dotenv==0.20.0
Werkzeug==2.0.3
WTForms==3.0.1
redis==4.2.2
gunicorn==20.1.0
